--- .idea/runConfigurations/main_readvcfs_test_annotate.xml
@@ -1,21 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="main_readvcfs_test_annotate" type="PythonConfigurationType" factoryName="Python">
-    <module name="gen-toolbox" />
-    <option name="INTERPRETER_OPTIONS" value="" />
-    <option name="PARENT_ENVS" value="true" />
-    <option name="SDK_HOME" value="" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
-    <option name="IS_MODULE_SDK" value="true" />
-    <option name="ADD_CONTENT_ROOTS" value="true" />
-    <option name="ADD_SOURCE_ROOTS" value="true" />
-    <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py" />
-    <option name="SCRIPT_NAME" value="$PROJECT_DIR$/main.py" />
-    <option name="PARAMETERS" value="readvcfs -f /home/villem/tests -d /home/villem/tests -g &quot;/mnt/sdb/phenotype-pseudon.txt&quot;" />
-    <option name="SHOW_COMMAND_LINE" value="true" />
-    <option name="EMULATE_TERMINAL" value="false" />
-    <option name="MODULE_MODE" value="false" />
-    <option name="REDIRECT_INPUT" value="false" />
-    <option name="INPUT_FILE" value="" />
-    <method v="2" />
-  </configuration>
-</component>
