+++ src/config/gene_config.json
@@ -1,0 +1,103 @@
+{
+    "intersect_genes": [
+        "AIP",
+        "APC",
+        "ATM",
+        "BAP1",
+        "BLM",
+        "BMPR1A",
+        "BRCA1",
+        "BRCA2",
+        "BRIP1",
+        "CDC73",
+        "CDH1",
+        "CDK4",
+        "CDKN2A",
+        "CEBPA",
+        "CHEK2",
+        "DDB2",
+        "DICER1",
+        "DIS3L2",
+        "EPCAM",
+        "ERCC2",
+        "ERCC3",
+        "ERCC4",
+        "ERCC5",
+        "FANCA",
+        "FANCB",
+        "FANCC",
+        "FANCD2",
+        "FANCE",
+        "FANCF",
+        "FANCG",
+        "FANCI",
+        "FANCL",
+        "FANCM",
+        "FH",
+        "FLCN",
+        "GATA2",
+        "GPC3",
+        "KIT",
+        "MAX",
+        "MEN1",
+        "MET",
+        "MLH1",
+        "MSH2",
+        "MSH6",
+        "MUTYH",
+        "NBN",
+        "NF1",
+        "NF2",
+        "NSD1",
+        "PALB2",
+        "PHOX2B",
+        "PMS2",
+        "PRKAR1A",
+        "PTCH1",
+        "PTEN",
+        "RAD51C",
+        "RAD51D",
+        "RB1",
+        "RECQL4",
+        "RET",
+        "RHBDF2",
+        "RUNX1",
+        "SDHAF2",
+        "SDHB",
+        "SDHC",
+        "SDHD",
+        "SLX4",
+        "SMAD4",
+        "SMARCB1",
+        "STK11",
+        "SUFU",
+        "TMEM127",
+        "TP53",
+        "TSC1",
+        "TSC2",
+        "VHL",
+        "WT1",
+        "XPA",
+        "XPC"
+    ],
+    "rv_genes": [
+        "ATP1A3",
+        "OTOF",
+        "DFNB59",
+        "OPA1",
+        "DIAPH3",
+        "ATP11A",
+        "CABP2",
+        "TMEM43"
+    ],
+    "neg_control_genes": [
+        "APC",
+        "BMPR1A",
+        "MSH2",
+        "MSH6",
+        "PTEN",
+        "BRCA1",
+        "BRCA2",
+        "WT1"
+    ]
+}
