+++ src/config/settings.py
@@ -1,0 +1,1 @@
+#(If needed in the future) For managing various configurations or settings.
