+++ src/config/vep_settings.json
@@ -1,0 +1,26 @@
+{
+    "command": [
+        "/mnt/sdb/projects/ensembl-vep/vep",
+        "--format", "vcf",
+        "__OUTPUT_FORMAT_FLAG__",
+        "--fasta", "/mnt/sdb/VEP/ref_fasta/ucsc.hg19.fasta",
+        "--dir_cache", "/mnt/sdb/VEP/VEP_cache/",
+        "--no_stats",
+        "--cache", "--offline",
+        "--assembly", "GRCh37",
+        "--cache_version", "108",
+        "--merged",
+        "--MAX_AF",
+        "--symbol",
+        "--fields", "IMPACT,SYMBOL,HGNC_ID,MAX_AF,MAX_AF_POPS",
+        "--pick",
+        "--canonical",
+        "--use_given_ref",
+        "--offline",
+        "-o", "STDOUT"
+    ],
+    "env": {
+        "PERL5LIB": "/vep_data/loftee"
+    },
+    "vep_json_schema": "Struct{IMPACT:String,SYMBOL:String,HGNC_ID:Int32,MAX_AF:Float64,MAX_AF_POPS:String,input:String}"
+}
