+++ src/utils/general/data_manipulation.py
@@ -1,0 +1,15 @@
+import os
+
+# Import a faster walker if it is installed
+try:
+    from scandir import walk
+except ImportError:
+    from os import walk
+
+import shutil
+import re
+
+def batcher(iterable, batch_size):
+    iterator = iter(iterable)
+    while batch := list(islice(iterator, batch_size)):
+        yield batch
